MozioTransport/requirements.txt
dj-database-url==0.5.0
Django==2.1
django-language-field==0.0.3
django-leaflet==0.24.0
django-money==0.14.3
djangorestframework==3.8.2
gunicorn==19.9.0
pep8==1.7.1
psycopg2==2.7.5
py-moneyed==0.7.0
pycodestyle==2.4.0
python-memcached==1.59
pytz==2018.5
rest-framework-cache==0.1
Shapely==1.6.4.post2
six==1.11.0
virtualenv==16.0.0
whitenoise==4.0
